requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy
PyYAML
pillow
opencv-python
scipy
tqdm
lpips
matplotlib
tensorboard
facenet-pytorch
pandas
kornia
scikit-learn
diffusers>=0.21.0
transformers>=4.30.0
torchattacks
